+++ src/shared/components/promiseDetail/PromisisDetail.css.ts
