+++ src/shared/components/promiseDetail/utils/getStatusSet.tsx
